+++ src/cares_wrap.cc
@@ -21,7 +21,7 @@
 
 #define CARES_STATICLIB
 #include "ares.h"
-#include "async-wrap-inl.h"
+#include "async_wrap-inl.h"
 #include "env-inl.h"
 #include "node.h"
 #include "req-wrap-inl.h"
