+++ src/js_stream.cc
@@ -1,6 +1,6 @@
 #include "js_stream.h"
 
-#include "async-wrap.h"
+#include "async_wrap.h"
 #include "env-inl.h"
 #include "node_buffer.h"
 #include "stream_base-inl.h"
