+++ src/node_http_parser.cc
@@ -22,7 +22,7 @@
 #include "node.h"
 #include "node_buffer.h"
 
-#include "async-wrap-inl.h"
+#include "async_wrap-inl.h"
 #include "env-inl.h"
 #include "http_parser.h"
 #include "stream_base-inl.h"
