+++ src/stream_base.h
@@ -4,7 +4,7 @@
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
 #include "env.h"
-#include "async-wrap.h"
+#include "async_wrap.h"
 #include "req-wrap-inl.h"
 #include "node.h"
 #include "util.h"
